Home

Islander.jl

Islander will be a nice component separation code someday!

Islander.Enmap

Islander.car

Islander.ellpad

Islander.inverse_gamma

Islander.util_get_cltt

Islander.Enmap — Type.

Map type, contains an AbstractArray and a WCS object, but behaves like the AbstractArray it contains for array operations.

It only implements the subset of Base.Array operations which are common on maps. You should work with the data directly using enmap_instance.data if you need additional Array functions.

source

Islander.car — Method.

Generate a plate carree WCS object. not implemented!

pixell python reference:

pos :
def car(pos, res=None, shape=None, rowmajor=False, ref=None):
	pos, res, shape, mid = validate(pos, res, shape, rowmajor)
	w = WCS(naxis=2)
	w.wcs.ctype = ["RA---CAR", "DEC--CAR"]
	w.wcs.crval = np.array([mid[0],0])
	if ref is "standard": ref = (0,0)
	return finalize(w, pos, res, shape, ref=ref)

source

Islander.ellpad — Method.

Utility function which prepends some zeros to an array. It makes a copy instead of modifying the input.

source

Islander.inverse_gamma — Method.

Draws an inverse gamma distribution given the sky map spectrum.

source

Islander.util_get_cltt — Function.

Utility function to load $C_{\ell}^{TT}$ when working in the notebooks.

source