Home

Islander.jl

Islander will be a nice component separation code someday!

Islander.Enmap

Islander.ell_binpow

Islander.ellpad

Islander.enfft!

Islander.enifft!

Islander.inverse_gamma

Islander.util_get_cltt

Islander.Enmap — Type.

Map type, contains an AbstractArray and a WCS object, but behaves like the AbstractArray it contains for array operations.

It only implements the subset of Base.Array operations which are common on maps. You should work with the data directly using enmap_instance.data if you need additional Array functions.

source

Islander.enfft! — Method.

Physically normalized enmap FFT

source

Islander.enifft! — Method.

Physically normalized enmap inverse FFT

source

Islander.ell_binpow — Function.

Quick and dirty binned, flat-sky power spectra.

source

Islander.ellpad — Method.

Utility function which prepends some zeros to an array. It makes a copy instead of modifying the input.

source

Islander.inverse_gamma — Method.

Draws an inverse gamma distribution given the sky map spectrum.

source

Islander.util_get_cltt — Function.

Utility function to load $C_{\ell}^{TT}$ when working in the notebooks.

source